move camera related things to camera.{h,c}
